refactor: Move astro content collection zod schema to types folder
