+++ src/types/content-collection-schemas.ts
